"aprggeometry fixes"

+++ AprgGeometry/AprgGeometry/src/TwoDimensions/Line.cpp
@@ -4,9 +4,9 @@
 #include <TwoDimensions/TwoDimensionsHelper.hpp>
 
 #include <algorithm>
+#include <cmath>
 #include <iterator>
 #include <set>
-
 using namespace std;
 
 namespace alba
@@ -38,11 +38,10 @@ Line::Line(Point const& first, Point const& second)
         m_cCoefficient = first.getY()*deltaX;
         break;
     case LineType::Vertical:
-        m_slope = INFINITY;
+        m_slope = static_cast<double>(INFINITY);
         m_yIntercept = 0;
         m_xIntercept = first.getX();
-        m_aCoefficient = deltaY;
-        m_bCoefficient = 0;
+        m_aCoefficient = deltaY;        m_bCoefficient = 0;
         m_cCoefficient = -first.getX()*deltaY;
         break;
     default:
@@ -70,11 +69,10 @@ Line::Line(double const aCoefficient, double const bCoefficient, double const cC
         m_xIntercept = 0;
         break;
     case LineType::Vertical:
-        m_slope = INFINITY;
+        m_slope = static_cast<double>(INFINITY);
         m_yIntercept = 0;
         m_xIntercept = -cCoefficient/aCoefficient;
-        break;
-    default:
+        break;    default:
         m_slope = -aCoefficient/bCoefficient;
         m_yIntercept = -cCoefficient/bCoefficient;
         m_xIntercept = -cCoefficient/aCoefficient;
