+++ AprgGeometry/AprgGeometry/src/TwoDimensions/Polygon.hpp
@@ -20,10 +20,9 @@ public:
     Polygon(std::initializer_list<Point> vertices)
     {
         static_assert(numberOfVertices>2, "The numberOfVertices is less than 2. Its not a polygon by definition.");
-        unsigned int limit = std::min(numberOfVertices, vertices.size());
+        unsigned int limit = std::min(numberOfVertices, static_cast<unsigned int>(vertices.size()));
         std::copy(vertices.begin(), vertices.begin()+limit, m_vertices.begin());
     }
-
     Lines getLines() const
     {
         Lines line;
