+++ AprgGeometry/AprgGeometry/src/TwoDimensions/Polynomial.hpp
@@ -21,11 +21,10 @@ public:
     Polynomial(std::initializer_list<double> coefficients)
     {
         static_assert(numberOfCoefficients>0, "The numberOfCoefficients is not more than 0. Its not a polynomial");
-        unsigned int limit = std::min(numberOfCoefficients, coefficients.size());
+        unsigned int limit = std::min(numberOfCoefficients, static_cast<unsigned int>(coefficients.size()));
         std::copy(coefficients.begin(), coefficients.begin()+limit, m_coefficients.begin());
         saveFirstDerivativeCoefficients();
     }
-
     Line getTangentLineAt(double const x) const
     {
         double slope = getSlopeAt(x);
